requirements.txt
telepot>=12.7
requests>=2.21.0
docker_py==1.10.6
